extract CALayer extention to a standalone file
